cmd/keylight: support relative integers in -b and -t
I can use "-b +5%" to increase the brightness a little, for example.
Use a flag.Value, which keeps the code manageable.

Note that, beyond the new feature with sign prefixes,
there's one behavior change.
The previous flags used "0" as the "unset" state,
meaning that both of the following commands would toggle the light:

	keylight
	keylight -b 0

I think this was simply an unintended side effect,
as the toggle code says "if neither flag is set".
Now, using that flag errors, as it's outside the valid range:

	$ keylight -b 0
	failed to set lights: brightness (0) out of range 3 <= x <= 100